+++ lib/src/application.dart
@@ -16,6 +16,9 @@ abstract class PteroApplication {
 
   @GET('/api/application/users')
   Future<FractalListMeta<ApplicationUser, PaginatedMeta>> getUsers({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+    //
     @Query('include') ApplicationUsersIncludes? include,
     @Query('sort') UsersSort? sort,
     @Query('page') int? page,
@@ -30,29 +33,44 @@ abstract class PteroApplication {
 
   @GET('/api/application/users/{id}')
   Future<Fractal<ApplicationUser>> getUser({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+    //
     @Path('id') required int id,
     @Query('include') ApplicationUsersIncludes? include,
   });
 
   @GET('/api/application/users/external/{externalId}')
   Future<Fractal<ApplicationUser>> getUserByExternalId({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+    //
     @Path('externalId') required String externalId,
     @Query('include') ApplicationUsersIncludes? include,
   });
 
   @POST('/api/application/users')
   Future<Fractal<ApplicationUser>> createUser({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+    //
     @Body() required CreateUserRequest request,
   });
 
   @PATCH('/api/application/users/{id}')
   Future<Fractal<ApplicationUser>> updateUser({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+    //
     @Path('id') required int id,
     @Body() required UpdateUserRequest request,
   });
 
   @DELETE('/api/application/users/{id}')
   Future<void> deleteUser({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+    //
     @Path('id') required int id,
   });
 
@@ -60,6 +78,9 @@ abstract class PteroApplication {
 
   @GET('/api/application/nodes')
   Future<FractalListMeta<Node, PaginatedMeta>> getNodes({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+    //
     @Query('include') NodeIncludes? include,
     @Query('sort') NodesSort? sort,
     @Query('page') int? page,
@@ -74,6 +95,9 @@ abstract class PteroApplication {
 
   @GET('/api/application/nodes/{id}')
   Future<Fractal<Node>> getNode({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+    //
     @Path('id') required int id,
     @Query('include') NodeIncludes? include,
   });
@@ -86,24 +110,36 @@ abstract class PteroApplication {
 
   @POST('/api/application/nodes')
   Future<FractalMeta<Node, ResourceMeta>> createNode({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+    //
     @Body() required CreateNodeRequest request,
   });
 
   //TODO: check that the response is in fact a [Node]
   @PATCH('/api/application/nodes/{id}')
   Future<FractalMeta<Node, ResourceMeta>> updateNode({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+    //
     @Path('id') required int id,
     @Body() required UpdateNodeRequest request,
   });
 
   @DELETE('/api/application/nodes/{id}')
   Future<void> deleteNode({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+    //
     @Path('id') required int id,
   });
 
   @GET('/api/application/nodes/{id}/allocations')
   Future<FractalListMeta<ApplicationAllocation, PaginatedMeta>>
       getNodeAllocations({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+    //
     @Path('id') required int id,
     @Query('include') AllocationIncludes? include,
     // TODO: AllocationSort