+++ lib/src/client.dart
@@ -26,6 +26,8 @@ abstract class DeprecatedPteroClient {
   @DELETE('/api/client/account/ssh-keys/{fingerprint}')
   Future<void> deleteSshKey1_8({
     @Path() required String fingerprint,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 }
 
@@ -39,6 +41,11 @@ abstract class PteroClient {
   static final removeNullResourcesInterceptor =
       RemoveNullResourcesInterceptor();
 
+  static final defaultInterceptors = [
+    HandleErrorInterceptor(),
+    RemoveNullResourcesInterceptor(),
+  ];
+
   /// Set up a Pterodactyl API Client in one go!
   /// [baseUrl] is the base URL of the Pterodactyl server.
   /// [key] is the API key of the Pterodactyl account.
@@ -48,7 +55,7 @@ abstract class PteroClient {
     String? key,
     Dio? dio,
     String userAgent = 'Dartactyl/v1',
-    bool enableErrorInterceptor = true,
+    bool addDefaultInterceptors = true,
   }) {
     dio ??= Dio();
 
@@ -62,10 +69,7 @@ abstract class PteroClient {
       ..headers[HttpHeaders.contentTypeHeader] = 'application/json'
       ..baseUrl = url;
 
-    dio.interceptors.addAll([
-      if (enableErrorInterceptor) HandleErrorInterceptor(),
-      RemoveNullResourcesInterceptor(),
-    ]);
+    if (addDefaultInterceptors) dio.interceptors.addAll(defaultInterceptors);
 
     return PteroClient(dio);
   }
@@ -99,36 +103,51 @@ abstract class PteroClient {
   PteroApplication get application => PteroApplication(dio, baseUrl: baseUrl);
 
   /// Alias of listVariables
-  Future<FractalListMeta<EggVariable, StartupMeta>> getStartup(
-          {required String serverId}) =>
-      listVariables(serverId: serverId);
+  Future<FractalListMeta<EggVariable, StartupMeta>> getStartup({
+    required String serverId,
+    CancelToken? cancelToken,
+    ProgressCallback? onProgress,
+  }) =>
+      listVariables(
+        serverId: serverId,
+        cancelToken: cancelToken,
+        onProgress: onProgress,
+      );
 
-  Future<Map<String, String>> listDockerImages(
-          {required String serverId}) async =>
-      (await getStartup(serverId: serverId)).dockerImages;
+  Future<Map<String, String>> listDockerImages({
+    required String serverId,
+    CancelToken? cancelToken,
+    ProgressCallback? onProgress,
+  }) async =>
+      (await getStartup(
+        serverId: serverId,
+        cancelToken: cancelToken,
+        onProgress: onProgress,
+      ))
+          .dockerImages;
 
   // /// Login to Pterodactyl using username and password.
   // ///
   // /// PUTS YOU INTO COOKIE MODE!!!
-  ///
-  /// [xsrfHeader] is the XSRF token
-  @POST('/auth/login')
-  Future<void> _login(
-    @Body() PteroLoginRequest credentials,
-    @Header("X-XSRF-TOKEN") String xsrfHeader,
-  );
+  // ///
+  // /// [xsrfHeader] is the XSRF token
+  // @POST('/auth/login')
+  // Future<void> _login(
+  //   @Body() PteroLoginRequest credentials,
+  //   @Header("X-XSRF-TOKEN") String xsrfHeader,
+  // );
 
-  /// Logout of Pterodactyl, ending your session.
-  ///
-  /// This will automatically remove your api token if
-  /// you did not disable the option in the client.
-  ///
-  /// You will need to add a cookie manager interceptor to make use of this
-  ///
-  /// Not fully implemented.
-  @experimental
-  @GET('/auth/logout')
-  Future<void> logout();
+  // /// Logout of Pterodactyl, ending your session.
+  // ///
+  // /// This will automatically remove your api token if
+  // /// you did not disable the option in the client.
+  // ///
+  // /// You will need to add a cookie manager interceptor to make use of this
+  // ///
+  // /// Not fully implemented.
+  // @experimental
+  // @GET('/auth/logout')
+  // Future<void> logout();
 
   /// Get a list of servers.
   ///
@@ -169,45 +188,64 @@ abstract class PteroClient {
     @Query('filter[description]') String? filterByDescription,
     // What servers to return by access type
     @Query('type') GetServersQueryType? type = GetServersQueryType.member,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Get system permissions
   @GET('/api/client/permissions')
-  Future<Fractal<SystemPermissions>> getSystemPermissions();
+  Future<Fractal<SystemPermissions>> getSystemPermissions({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+  });
 
   /*       Account       */
 
   /// Get account information.
   @GET('/api/client/account')
-  Future<Fractal<Account>> getAccountInfo();
+  Future<Fractal<Account>> getAccountInfo({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+  });
 
   /// Get two factor authentication image.
   @GET('/api/client/account/two-factor')
-  Future<Fractal<TwoFactorImage>> getTwoFactor();
+  Future<Fractal<TwoFactorImage>> getTwoFactor({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+  });
 
   /// Enable two factor authentication.
   @POST('/api/client/account/two-factor')
   Future<Fractal<RecoveryTokens>> enableTwoFactor(
-    @Body() TwoFactorCode code,
-  );
+    @Body() TwoFactorCode code, {
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+  });
 
   /// Disable two factor authentication.
   @DELETE('/api/client/account/two-factor')
   Future<void> disableTwoFactor(
-    @Body() DisableTwoFactor data,
-  );
+    @Body() DisableTwoFactor data, {
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+  });
 
   /// Update your [Account] account email address.
   @PUT('/api/client/account/email')
   Future<void> updateEmail(
-    @Body() UpdateEmail data,
-  );
+    @Body() UpdateEmail data, {
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+  });
 
   /// Update your [Account] account password.
   @PUT('/api/client/account/password')
   Future<void> updatePassword(
-    @Body() UpdatePassword data,
-  );
+    @Body() UpdatePassword data, {
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+  });
 
   @GET('/api/client/account/activity')
   Future<FractalListMeta<ActivityLog, PaginatedMeta>> getAccountActivity({
@@ -221,41 +259,57 @@ abstract class PteroClient {
     @Query('filter[event]') String? filterByEvent,
     // Sort
     @Query('sort') ActivityLogSort? sort,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Get all current [ApiKey]s on your account.
   /// Keys are shortened to the first x characters.
   @GET('/api/client/account/api-keys')
-  Future<FractalList<ApiKey>> listApiKeys();
+  Future<FractalList<ApiKey>> listApiKeys({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+  });
 
   /// Create a new [ApiKey] on your account.
   /// This is the only time you will ever get the full key.
   @POST('/api/client/account/api-keys')
   Future<FractalMeta<ApiKey, ApiKeyMeta>> createApiKey(
-    @Body() CreateApiKey data,
-  );
+    @Body() CreateApiKey data, {
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+  });
 
   /// Delete an [ApiKey] on your account.
   @DELETE('/api/client/account/api-keys/{apiKeyId}')
   Future<void> deleteApiKey({
     @Path() required String apiKeyId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// List all [SshKey]s on your account.
   @GET('/api/client/account/ssh-keys')
-  Future<FractalList<SshKey>> listSshKeys();
+  Future<FractalList<SshKey>> listSshKeys({
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+  });
 
   /// Create a new [SshKey] on your account.
   @POST('/api/client/account/ssh-keys')
   Future<Fractal<SshKey>> createSshKey(
-    @Body() CreateSshKey data,
-  );
+    @Body() CreateSshKey data, {
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+  });
 
   /// Delete an [SshKey] on your account. (1.9+)
   @POST('/api/client/account/ssh-keys/remove')
   Future<void> deleteSshKey(
-    @Body() DeleteSSHKey body,
-  );
+    @Body() DeleteSSHKey body, {
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
+  });
 
   // '/api/client/servers/{server}'
 
@@ -266,18 +320,24 @@ abstract class PteroClient {
   Future<FractalMeta<Server, ServerMeta>> getServerDetails({
     @Path() required String serverId,
     @Query('include') ServerIncludes? include,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Get the [Server]'s [WebsocketDetails].
   @GET('/api/client/servers/{serverId}/websocket')
   Future<PteroData<WebsocketDetails>> getServerWebsocket({
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Get the [Server]'s current [Stats].
   @GET('/api/client/servers/{serverId}/resources')
   Future<Fractal<Stats>> getServerResources({
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   @GET('/api/client/servers/{serverId}/activity')
@@ -293,6 +353,8 @@ abstract class PteroClient {
     @Query('filter[event]') String? filterByEvent,
     // Sort
     @Query('sort') ActivityLogSort? sort,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Send a command to the [Server].
@@ -300,6 +362,8 @@ abstract class PteroClient {
   Future<void> sendServerCommand(
     @Body() SendServerCommand data, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// send a Power [Signal] to the [Server].
@@ -307,6 +371,8 @@ abstract class PteroClient {
   Future<void> sendServerPowerAction(
     @Body() Signal signal, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   // Databases
@@ -324,6 +390,8 @@ abstract class PteroClient {
     @Query('per_page') int? perPage,
     // filters
     // ??
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Create a new database on the server
@@ -331,6 +399,8 @@ abstract class PteroClient {
   Future<Fractal<ServerDatabase>> createServerDatabase(
     @Body() CreateServerDatabase data, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Delete a [ServerDatabase]
@@ -338,6 +408,8 @@ abstract class PteroClient {
   Future<void> deleteDatabase({
     @Path() required String serverId,
     @Path() required String databaseId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// TODO: on [rotateDatabasePassword]
@@ -345,6 +417,8 @@ abstract class PteroClient {
   Future<Fractal<ServerDatabase>> rotateDatabasePassword({
     @Path() required String serverId,
     @Path() required String databaseId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   // Files
@@ -356,6 +430,8 @@ abstract class PteroClient {
   Future<FractalList<FileObject>> listFiles({
     @Path() required String serverId,
     @Query('directory', encoded: true) required String directory,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Get a [file]'s contents from the [Server]
@@ -365,6 +441,8 @@ abstract class PteroClient {
   Future<String?> getFileContents({
     @Path() required String serverId,
     @Query('file', encoded: true) required String file,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Download a [file] from the [Server]
@@ -374,6 +452,8 @@ abstract class PteroClient {
   Future<Fractal<SignedUrl>> downloadFile({
     @Path() required String serverId,
     @Query('file', encoded: true) required String file,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Rename a file on the [Server]
@@ -381,6 +461,8 @@ abstract class PteroClient {
   Future<void> renameFile(
     @Body() FileBodyListFromTo rename, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Make a copy of a file on the [Server]
@@ -388,6 +470,8 @@ abstract class PteroClient {
   Future<void> makeFileCopy(
     @Body() MakeFileCopy data, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Write a [file] to the [Server]
@@ -401,6 +485,8 @@ abstract class PteroClient {
     @Body() String rawContents, {
     @Path() required String serverId,
     @Query('file', encoded: true) required String file,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Compress a file into an archive (eg. zip) on the [Server]
@@ -408,6 +494,8 @@ abstract class PteroClient {
   Future<Fractal<FileObject>> compressFile(
     @Body() FileBodyListString data, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Decompress an archive (eg. zip) on the [Server]
@@ -415,6 +503,8 @@ abstract class PteroClient {
   Future<void> decompressFile(
     @Body() FileBody data, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Delete one or more files on the [Server]
@@ -422,6 +512,8 @@ abstract class PteroClient {
   Future<void> deleteFiles(
     @Body() FileBodyListString data, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Creates the specified folder in the specified directory
@@ -429,6 +521,8 @@ abstract class PteroClient {
   Future<void> createFolder(
     @Body() FolderBody data, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Changes the permissions of a file or folder on the [Server]
@@ -436,6 +530,8 @@ abstract class PteroClient {
   Future<void> chmodFile(
     @Body() ChmodFileBody data, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Download a file from a remote url to the [Server] directly
@@ -443,12 +539,16 @@ abstract class PteroClient {
   Future<void> pullFile(
     @Body() PullFileBody data, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Returns a [SignedUrl] used to upload files to the [Server] using POST
   @GET('/api/client/servers/{serverId}/files/upload')
   Future<Fractal<SignedUrl>> getFileUploadUrl({
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   // Schedules
@@ -456,6 +556,8 @@ abstract class PteroClient {
   @GET('/api/client/servers/{serverId}/schedules')
   Future<FractalList<ServerSchedule>> listSchedules({
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Create a [ServerSchedule] on the [Server]
@@ -463,6 +565,8 @@ abstract class PteroClient {
   Future<Fractal<ServerSchedule>> createSchedule(
     @Body() RequestSchedule scheduleData, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Get a [ServerSchedule]'s details from the [Server]
@@ -470,6 +574,8 @@ abstract class PteroClient {
   Future<Fractal<ServerSchedule>> getScheduleDetails({
     @Path() required String serverId,
     @Path() required int scheduleId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Update a [ServerSchedule] on the [Server]
@@ -478,6 +584,8 @@ abstract class PteroClient {
     @Body() RequestSchedule scheduleData, {
     @Path() required String serverId,
     @Path() required int scheduleId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Delete a [ServerSchedule] from the [Server]
@@ -485,6 +593,8 @@ abstract class PteroClient {
   Future<void> deleteSchedule({
     @Path() required String serverId,
     @Path() required int scheduleId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Create a scheduled [Task] on a [ServerSchedule]
@@ -493,6 +603,8 @@ abstract class PteroClient {
     @Body() Task taskData, {
     @Path() required String serverId,
     @Path() required int scheduleId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Update a scheduled [Task] on a [ServerSchedule]
@@ -502,6 +614,8 @@ abstract class PteroClient {
     @Path() required String serverId,
     @Path() required int scheduleId,
     @Path() required int taskId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Delete a scheduled [Task] on a [ServerSchedule]
@@ -511,6 +625,8 @@ abstract class PteroClient {
     @Path() required String serverId,
     @Path() required int scheduleId,
     @Path() required int taskId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   // Network
@@ -518,12 +634,16 @@ abstract class PteroClient {
   @GET('/api/client/servers/{serverId}/network')
   Future<FractalList<Allocation>> listAllocations({
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Automatically assign an allocation on the [Server]
   @POST('/api/client/servers/{serverId}/network')
   Future<Fractal<Allocation>> autoAssignAllocation({
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Set the allocation note for an [Allocation] on the [Server]
@@ -532,6 +652,8 @@ abstract class PteroClient {
     @Body() AllocationNote note, {
     @Path() required String serverId,
     @Path() required int allocationId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Set an [Allocation] as the primary allocation on [Server]
@@ -539,6 +661,8 @@ abstract class PteroClient {
   Future<Fractal<Allocation>> setPrimaryAllocation({
     @Path() required String serverId,
     @Path() required int allocationId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Unassign an [Allocation] from [Server]
@@ -546,6 +670,8 @@ abstract class PteroClient {
   Future<Fractal<Allocation>> unassignAllocation({
     @Path() required String serverId,
     @Path() required int allocationId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /*                                  Users                                  */
@@ -554,6 +680,8 @@ abstract class PteroClient {
   @GET('/api/client/servers/{serverId}/users')
   Future<FractalList<ServerSubuser>> listSubusers({
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Create [ServerSubuser] on the [Server]
@@ -561,6 +689,8 @@ abstract class PteroClient {
   Future<Fractal<ServerSubuser>> createSubuser(
     @Body() Subuser subuserData, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Get a [ServerSubuser]'s details on the [Server]
@@ -568,6 +698,8 @@ abstract class PteroClient {
   Future<Fractal<ServerSubuser>> getSubuserDetails({
     @Path() required String serverId,
     @Path() required String subuserId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Update a [ServerSubuser] on the [Server]
@@ -576,6 +708,8 @@ abstract class PteroClient {
     @Body() SubuserPermissions subuserData, {
     @Path() required String serverId,
     @Path() required String subuserId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Delete a [ServerSubuser] from the [Server]
@@ -583,6 +717,8 @@ abstract class PteroClient {
   Future<void> deleteSubuser({
     @Path() required String serverId,
     @Path() required String subuserId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   // Backups
@@ -594,6 +730,8 @@ abstract class PteroClient {
     // pagination
     @Query('page') int? page,
     @Query('per_page') int? perPage,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Create a backup on the [Server]
@@ -601,6 +739,8 @@ abstract class PteroClient {
   Future<Fractal<Backup>> createBackup(
     @Body() CreateBackup backupData, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Get information about a [Backup] from the [Server]
@@ -608,6 +748,8 @@ abstract class PteroClient {
   Future<Fractal<Backup>> getBackupDetails({
     @Path() required String serverId,
     @Path() required String backupId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Lock a [Backup] to protect it from automated or accedental deletions
@@ -616,6 +758,8 @@ abstract class PteroClient {
   Future<Fractal<Backup>> toggleBackupLock({
     @Path() required String serverId,
     @Path() required String backupId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Generate download url for a [Backup] from  the [Server]
@@ -623,6 +767,8 @@ abstract class PteroClient {
   Future<Fractal<SignedUrl>> downloadBackup({
     @Path() required String serverId,
     @Path() required String backupId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Delete a [Backup] from the [Server]
@@ -630,6 +776,8 @@ abstract class PteroClient {
   Future<void> deleteBackup({
     @Path() required String serverId,
     @Path() required String backupId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Restore a [Backup] to the [Server]
@@ -638,6 +786,8 @@ abstract class PteroClient {
     @Body() RestoreBackup body, {
     @Path() required String serverId,
     @Path() required String backupId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   // Startup
@@ -646,6 +796,8 @@ abstract class PteroClient {
   @GET('/api/client/servers/{serverId}/startup')
   Future<FractalListMeta<EggVariable, StartupMeta>> listVariables({
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Update the [Server] startup variable with the contents of [variable]
@@ -653,6 +805,8 @@ abstract class PteroClient {
   Future<Fractal<EggVariable>> updateVariable(
     @Body() KeyValue variable, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   // Settings
@@ -661,12 +815,16 @@ abstract class PteroClient {
   Future<void> renameServer(
     @Body() Rename rename, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Reinstall the [Server]
   @POST('/api/client/servers/{serverId}/settings/reinstall')
   Future<void> reinstallServer({
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 
   /// Update the [Server] docker image to [dockerImage]
@@ -674,5 +832,7 @@ abstract class PteroClient {
   Future<void> updateDockerImage(
     @Body() UpdateImage dockerImage, {
     @Path() required String serverId,
+    @CancelRequest() CancelToken? cancelToken,
+    @SendProgress() ProgressCallback? onProgress,
   });
 }